Fix Redshift profiler chart hour column type
date_part(hour, start_time) returns DOUBLE PRECISION in Redshift, but the DDL columns are BIGINT. Cast to ::int in charts 6/8/9 so hour is an integer 0-23 matching the declared column type.

+++ src/databricks/labs/lakebridge/resources/assessments/redshift/sql/9_chart_cpu_consumption_by_hour_and_query_type.sql
@@ -8,7 +8,7 @@ with query_metrics as (select user_id
 select 'chart_cpu_consumption_by_hour_and_query_type' set_name
       ,query_type
       ,sum(run_time_ms)::double precision as sum_cpu_time
-      ,date_part(hour,start_time) as hour
+      ,date_part(hour,start_time)::int as hour
   from query_overview
  where run_time_ms > 0
  group by 1,2,4
